android test

+++ appAndroid/src/main/kotlin/hummel/dao/impl/FileDaoImpl.kt
@@ -1,13 +1,13 @@
 package hummel.dao.impl
 
-import android.content.Context
+import hummel.bean.BotData
 import hummel.dao.FileDao
 import hummel.utils.random
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
 
-class FileDaoImpl(private var context: Context) : FileDao {
+class FileDaoImpl(private val botData: BotData) : FileDao {
 	override fun createFile(filePath: String) {
 		val file = getFile(filePath)
 		if (!file.exists()) {
@@ -36,45 +36,58 @@ class FileDaoImpl(private var context: Context) : FileDao {
 		}
 	}
 
-	override fun getFile(filePath: String): File = File(context.filesDir, filePath)
+	override fun getFile(filePath: String): File = File(botData.context as String, filePath)
 
-	override fun getFolder(folderPath: String?): File =
-		folderPath?.let { File(context.filesDir, it) } ?: File(context.filesDir.path)
+	override fun getFolder(folderPath: String): File = File(botData.context as String, folderPath)
 
 	override fun readFromFile(filePath: String): ByteArray {
 		var byteArray: ByteArray
 		val file = getFile(filePath)
-		FileInputStream(file).use {
-			byteArray = ByteArray(it.available())
-			it.read(byteArray)
+		if (file.exists()) {
+			FileInputStream(file).use {
+				byteArray = ByteArray(it.available())
+				it.read(byteArray)
+			}
+		} else {
+			throw Exception("File doesn't exist!")
 		}
 		return byteArray
 	}
 
 	override fun writeToFile(filePath: String, byteArray: ByteArray) {
 		val file = getFile(filePath)
-		FileOutputStream(file).use {
-			it.write(byteArray)
+		if (file.exists()) {
+			FileOutputStream(file).use {
+				it.write(byteArray)
+			}
+		} else {
+			throw Exception("File doesn't exist!")
 		}
 	}
 
 	override fun appendToFile(filePath: String, byteArray: ByteArray) {
 		val file = getFile(filePath)
-		FileOutputStream(file, true).use {
-			it.write(byteArray)
+		if (file.exists()) {
+			FileOutputStream(file, true).use {
+				it.write(byteArray)
+			}
+		} else {
+			throw Exception("File doesn't exist!")
 		}
 	}
 
 	override fun getRandomLine(filePath: String): String? {
 		val file = getFile(filePath)
-		val lines = file.readLines()
-		if (lines.isNotEmpty()) {
-			val randomLine = lines[random.nextInt(lines.size)]
-			if (randomLine.isNotEmpty()) {
-				return randomLine
+		if (file.exists()) {
+			val lines = file.readLines()
+			if (lines.isNotEmpty()) {
+				val randomLine = lines[random.nextInt(lines.size)]
+				if (randomLine.isNotEmpty()) {
+					return randomLine
+				}
 			}
 			return null
 		}
-		return null
+		throw Exception("File doesn't exist!")
 	}
 }
